Move patching to the base class
